カーソルと一緒にノーツも進むように変更
